backend/requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=23.0
Pillow